기존 Header.jsx 파일 삭제 후, 반응형 UI를 적용한 새로운 Header.jsx로 수정
